chore: extract svm-type crate
