+++ addons/svm/core/Cargo.toml
@@ -11,6 +11,7 @@ categories = { workspace = true }
 [dependencies]
 # txtx-addon-kit = { version = "0.2.2", default-features = false }
 txtx-addon-kit = { workspace = true, default-features = false }
+txtx-addon-network-svm-types =  { path = "../types" }
 lazy_static = "1.4.0"
 serde_json = "1.0.113"
 serde = "1"
@@ -21,14 +22,13 @@ solana-sdk = "2.2.1"
 solana-client = "2.2.1"
 spl-associated-token-account = "6.0.0"
 spl-token = "7.0.0"
-anchor-lang-idl = { version = "0.1.2", features = ["convert"] }
 solana_idl = "0.2.0"
 borsh = "1.5.1"
 tiny-bip39 = "0.8.2"
 convert_case = "0.6.0"
 
 [dev-dependencies]
-txtx-test-utils = { path = "../../crates/txtx-test-utils" }
+txtx-test-utils = { path = "../../../crates/txtx-test-utils" }
 
 [features]
 default = [
