+++ addons/svm/core/examples/hellosol/.gitignore
