+++ addons/svm/core/examples/hellosol/Cargo.toml
