+++ addons/svm/core/examples/hellosol/programs/hellosol/Cargo.toml
