+++ addons/svm/core/examples/hellosol/programs/hellosol/Xargo.toml
