+++ addons/svm/core/examples/hellosol/programs/hellosol2/Cargo.toml
