+++ addons/svm/core/examples/hellosol/programs/hellosol2/Xargo.toml
