+++ addons/svm/core/src/codec/idl.rs
@@ -1,18 +1,18 @@
 use std::{path::PathBuf, str::FromStr};
 
-use crate::typing::{SvmValue, SVM_PUBKEY};
+use crate::typing::anchor as anchor_lang_idl;
+use crate::typing::SvmValue;
 use anchor_lang_idl::types::{
     Idl, IdlArrayLen, IdlDefinedFields, IdlGenericArg, IdlInstruction, IdlInstructionAccount,
     IdlInstructionAccounts, IdlType, IdlTypeDef, IdlTypeDefGeneric, IdlTypeDefTy,
 };
-use serde_json::Value as JsonValue;
 use solana_idl::Idl as ClassicIdl;
-use solana_sdk::{bs58, pubkey::Pubkey};
+use solana_sdk::pubkey::Pubkey;
 use spl_token::solana_program;
 use txtx_addon_kit::{
     helpers::fs::FileLocation,
     indexmap::IndexMap,
-    types::types::{ObjectType, Type, Value},
+    types::types::{ObjectType, Value},
 };
 
 pub struct IdlRef {
@@ -460,64 +460,6 @@ pub fn encode_value_to_idl_type(
             }
         }
         t => return Err(format!("IDL type {:?} is not yet supported", t)),
-    }
-}
-
-pub fn get_expected_field_type_from_idl_type_def_ty(
-    field_name: &str,
-    idl_type_def_ty: &IdlTypeDefTy,
-) -> Result<Type, String> {
-    let ty = match idl_type_def_ty {
-        IdlTypeDefTy::Struct { fields } => {
-            let ty = if let Some(fields) = fields {
-                let field = match fields {
-                    IdlDefinedFields::Named(idl_fields) => idl_fields
-                        .iter()
-                        .find(|f| f.name == field_name)
-                        .ok_or(format!("unable to find field '{}' in struct", field_name))?,
-                    IdlDefinedFields::Tuple(idl_types) => {
-                        return Err("cannot find field by name for tuple type".to_string())
-                    }
-                };
-                field.ty.clone()
-            } else {
-                return Err(format!("unable to find field '{}' in struct", field_name));
-            };
-            ty
-        }
-        IdlTypeDefTy::Enum { variants } => todo!(),
-        IdlTypeDefTy::Type { alias } => todo!(),
-    };
-
-    Ok(idl_type_to_txtx_type(ty))
-}
-
-pub fn idl_type_to_txtx_type(idl_type: IdlType) -> Type {
-    match idl_type {
-        IdlType::Bool => Type::bool(),
-        IdlType::U8 => Type::integer(),
-        IdlType::I8 => Type::integer(),
-        IdlType::U16 => Type::integer(),
-        IdlType::I16 => Type::integer(),
-        IdlType::U32 => Type::integer(),
-        IdlType::I32 => Type::integer(),
-        IdlType::U64 => Type::integer(),
-        IdlType::I64 => Type::integer(),
-        IdlType::I128 => Type::integer(),
-        IdlType::F32 => Type::float(),
-        IdlType::F64 => Type::float(),
-        IdlType::U128 => todo!(),
-        IdlType::U256 => todo!(),
-        IdlType::I256 => todo!(),
-        IdlType::Bytes => Type::buffer(),
-        IdlType::String => Type::string(),
-        IdlType::Pubkey => Type::addon(SVM_PUBKEY),
-        IdlType::Option(idl_type) => idl_type_to_txtx_type(*idl_type),
-        IdlType::Vec(idl_type) => Type::array(idl_type_to_txtx_type(*idl_type)),
-        IdlType::Array(idl_type, ..) => Type::array(idl_type_to_txtx_type(*idl_type)),
-        IdlType::Defined { .. } => todo!(),
-        IdlType::Generic(_) => todo!(),
-        _ => todo!(),
     }
 }
