+++ addons/svm/core/src/codec/native.rs
