+++ addons/svm/core/src/commands/process_instructions.rs
