+++ addons/svm/core/src/constants.rs
@@ -61,7 +61,6 @@ pub const IS_FUNDING_RECIPIENT: &str = "is_funding_recipient";
 // Subgraph keys
 pub const BLOCK_HEIGHT: &str = "block_height";
 pub const EVENT: &str = "event";
-pub const FIELD: &str = "field";
 pub const SUBGRAPH_NAME: &str = "subgraph_name";
 pub const SUBGRAPH_REQUEST: &str = "subgraph_request";
 pub const SUBGRAPH_URL: &str = "subgraph_url";
