+++ addons/svm/core/src/lib.rs
@@ -11,8 +11,12 @@ pub mod functions;
 pub mod rpc;
 mod signers;
 pub mod templates;
-pub mod typing;
+pub mod utils;
 pub use solana_sdk::pubkey::Pubkey;
+
+pub mod typing {
+    pub use txtx_addon_network_svm_types::*;
+}
 
 use constants::NAMESPACE;
 use txtx_addon_kit::{
