+++ addons/svm/types/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "txtx-addon-network-svm-types"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+solana-pubkey = { version = "2.2.1", features = ["serde"]}
+solana-transaction-context = { version = "2.2.1", features = ["serde"]}
+solana-clock = { version = "2.2.1", features = ["serde"]}
+solana-blake3-hasher = { version = "2.2.1", features = ["serde"]}
+solana-epoch-info = { version = "2.2.1", features = ["serde"]}
+solana-message = { version = "2.2.1", features = ["serde"]}
+solana-keypair = { version = "2.2.1" }
+solana-signature = { version = "2.2.1", features = ["serde"]}
+solana-transaction = { version = "2.2.1", features = ["serde", "bincode"]}
+solana-transaction-error = { version = "2.2.1", features = ["serde"]}
+anchor-lang-idl = { version = "0.1.2", features = ["convert"] }
+txtx-addon-kit = { workspace = true, default-features = false }
+lazy_static = "1.4.0"
+serde = "1"
+serde_json = "1"
+serde_derive = "1"
